Move Component and Context to a separate file
